LayoutTests/http/tests/security/filesystem-iframe-from-remote.html

<html> <head>
    <script>
        function done() {
            var pageLocation = "filesystem:http://localhost:8000/temporary/page.html";
            var localIframeElement = document.createElement("iframe");
            localIframeElement.setAttribute("id", "myFrame");
            localIframeElement.setAttribute("src", pageLocation);
            document.body.appendChild(localIframeElement);

            var result = document.getElementById("result");
            var myFrameDocument = document.getElementById("myFrame").contentDocument;
            if (myFrameDocument)
                result.innerHTML = "Test Failed: Local page remotely loaded into iFrame.";
            else
                result.innerHTML = "Test Passed. Local page not remotely loaded into iFrame.";

            if (window.testRunner)
                testRunner.notifyDone();
        }

        function runTest() {
            if (window.testRunner) {
                testRunner.dumpAsText();
                testRunner.waitUntilDone();
            }
            // Need to wait for the iframe to do its async work.
            setTimeout(done, 1000);
        }

    </script>
</head>
<body>
    <div id="div0">
        This test is to see if a remote file can include a local page via a filesystem: URI in an iFrame.
    </div>
    <br>
    <div id="result">
        Test not run correctly.
    </div>
    <iframe src="http://localhost:8000/security/resources/create-filesystem-file.html" onload="runTest()"></iframe>
</body>
</html>
