LayoutTests/http/tests/security/filesystem-iframe-from-remote-expected.txt

CONSOLE MESSAGE: line 11: filesystem:http://localhost:8000/temporary/page.html created.
CONSOLE MESSAGE: Not allowed to load local resource: filesystem:http://localhost:8000/temporary/page.html
This test is to see if a remote file can include a local page via a filesystem: URI in an iFrame.

Test Passed. Local page not remotely loaded into iFrame.
  
